+++ GROBS/Models/ord_fanliViewModel.cs
@@ -11,9 +11,9 @@ namespace GROBS.Models
         [Display(Name = "客户")]
         public int? KHID { get; set; }
         [Display(Name = "总额")]
-        public float? Zonge { get; set; }
+        public decimal? Zonge { get; set; }
         [Display(Name = "可用")]
-        public float? Keyong { get; set; }
+        public decimal? Keyong { get; set; }
         [Display(Name = "机动1")]
         public string Col1 { get; set; }
         [Display(Name = "制单日期")]
